novice/04-05/requirements.txt
Flask
python-dotenv
psycopg2-binary
SQLAlchemy 